pyproject.toml
[tool.poetry]
name = "forms"
version = "0.1.0"
description = ""
authors = ["Roberta Ferreira Aguilar <roberta_rfa@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
tabulate = "^0.9.0"
fpdf = "^1.7.2"
python-docx = "^1.1.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
